+++ CHANGES.rst
@@ -10,9 +10,6 @@ Incompatibilities:
 
 New:
 
-- Add new tests for sitemap.xml.gz (it is currently not listing any content)
-  [djowett]
-
 - *add item here*
 
 Fixes:
@@ -28,6 +25,9 @@ Fixes:
 
 - Update Traditional Chinese Translations.
   [l34marr]
+
+- Fix for broken sitemap.xml.gz (which was not listing any content #209)
+  [djowett]
 
 - *add item here*